Move mouse utils to annotator folder
